# api/openapi.yaml
openapi: 3.0.0
info:
  title: HR Recruiter API
  version: 1.0.0
  description: |
    Полная OpenAPI спецификация для платформы автоматизации.
    Включает все основные сущности, бизнес-операции, заглушки для AI и медиа, а также примеры для быстрой интеграции.
servers:
  - url: http://localhost:8080/api/v1
    description: Local server
  - url: https://api.hr-recruiter.com/api/v1
    description: Production server
...
paths:
...
  /ai/generate-position:
    post:
      operationId: generatePosition
      tags:
        - AI
      summary: Генерировать вакансию с помощью AI
      description: |
        Генерирует структуру вакансии на основе описания пользователя.
        AI возвращает минимальный набор полей (title, description, topics, level, questions).
        Остальные поля (status, language, settings и т.д.) устанавливаются системой по умолчанию.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionAiGenerationRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionAiGenerationResponse'
        '400':
          description: Bad Request - некорректные входные данные
        '500':
          description: Internal Server Error - ошибка AI или системы
...
components:
  schemas:
...
    PositionAiGenerationRequest:
      type: object
      description: "Запрос на генерацию вакансии с помощью AI"
      properties:
        description:
          type: string
          description: "Описание вакансии от пользователя"
          example: "Нужен Java разработчик для разработки микросервисов"
        questionsCount:
          type: integer
          description: "Количество вопросов для собеседования"
          example: 5
          minimum: 1
          maximum: 20
        questionType:
          type: string
          description: "Тип вопросов (hard, soft, mixed)"
          example: "hard"
      required: [description]
    PositionAiGenerationResponse:
      type: object
      description: "Ответ AI с минимальной структурой вакансии"
      properties:
        title:
          type: string
          description: "Название вакансии"
          example: "Java Backend Developer"
        description:
          type: string
          description: "Описание вакансии"
          example: "Разработка микросервисов на Java с использованием Spring Framework"
        topics:
          type: array
          items:
            type: string
          description: "Ключевые темы/навыки для собеседования"
          example: ["Java Core", "Spring Framework", "Microservices", "Database Design"]
        level:
          type: string
          enum: [junior, middle, senior, lead]
          description: "Уровень позиции"
          example: "middle"
        questions:
          type: array
          items:
            $ref: '#/components/schemas/PositionAiQuestion'
          description: "Список вопросов для собеседования"
      required: [title, topics, level, questions]
    PositionAiQuestion:
      type: object
      description: "Вопрос для собеседования (AI-генерация)"
      properties:
        text:
          type: string
          description: "Текст вопроса"
          example: "Расскажите о принципах SOLID"
        type:
          type: string
          enum: [text, audio, choice]
          description: "Тип вопроса"
          example: "text"
        order:
          type: integer
          description: "Порядок вопроса"
          example: 1
      required: [text, type, order]
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
security:
  - basicAuth: []
...